pyproject.toml
[tool.poetry]
name = "automatic-tweet-generator"
version = "0.1.0"
description = "Generated quoted image tweets using web scraping and api calls"
authors = ["bradleybares <bradleygbares@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.27.1"
requests-html = "^0.10.0"
Pillow = "^9.0.1"
tweepy = "^4.6.0"
loguru = "^0.6.0"
schedule = "^1.1.0"
pandas = "^1.4.1"

[tool.poetry.dev-dependencies]
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
